revise

+++ src/Algorithm/AlgorithmBase/AlgorithmBase.cpp
@@ -51,7 +51,7 @@ void AlgorithmBase::base_next_time_slot(){
         if(requests[reqno].get_throughput() == 0){
             req_success_ratio = 0;
         }else{
-            req_success_ratio = requests[reqno].get_throughput() / requests[reqno].get_send_limit();
+            req_success_ratio = (double)requests[reqno].get_throughput() / (double)requests[reqno].get_send_limit();
             if(max_over_ratio <= req_success_ratio){
                 max_over_ratio = req_success_ratio;
             }
